Changed Niling Fiber again.
Instead of chain-detonating across a preset radius until it runs out of power, it now detonates once for the same amount of damage, scaling the radius of the explosion based on the amount of detonations it would previously have triggered.

(... That made no sense whatsoever. Here's hoping my localization will do a better job of explaining it.)